Fix the one narrowing bounds check this branch added
CodeQL raised a new high-severity alert on the pull request: the ghost sheet
narrowed `ghost_form`, an int64 read out of SQLite, to `int` before comparing
it against the length of the ghost-form ladder. The earlier sweep fixed the
nineteen pre-existing cases of this and missed the one the branch itself had
just introduced.

Same fix as the rest: widen the length rather than narrow the index.

+++ go_core/internal/game/death_qi.go
@@ -406,7 +406,7 @@ func ghostStatusQuery(conn *storage.Conn, catalog worlddata.Catalog, userID int6
 	tm := cultivationTimeModifiers(gameMinute, c.SpiritualRoot)
 	form := ghostFormAt(catalog, body.GhostForm)
 	next := map[string]any(nil)
-	if int(body.GhostForm)+1 < len(catalog.DeathQi.GhostForms) {
+	if body.GhostForm+1 < int64(len(catalog.DeathQi.GhostForms)) {
 		step := catalog.DeathQi.GhostForms[body.GhostForm+1]
 		next = map[string]any{
 			"name": step.Name, "corruption": step.Corruption, "min_realm_index": step.MinRealmIndex,
